+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Census3 error typings.
+- Census3 anonymous censuses.
 
 ## [0.1.1] - 2023-08-14